Moved the stylesheet to a different file so that our codebase is less messy.
